[Python] Fix BinaryView.get_modification failing when passing a length
Fixes https://github.com/Vector35/binaryninja-api/issues/8046

+++ python/binaryview.py
@@ -5221,7 +5221,7 @@ class BinaryView:
 		"""
 		if length is None:
 			return [ModificationStatus(core.BNGetModification(self.handle, addr))]
-		data = (ctypes.c_int * length)()
+		data = (ctypes.c_ubyte * length)()
 		length = core.BNGetModificationArray(self.handle, addr, data, length)
 		return [ModificationStatus(a) for a in data[:length]]
